Making things changeable
Made it so that the elf parser returns a little 64-bit array. This way
the address and the offset within the file where the kernel starts and
the address where it needs to be mapped is returned. The little 64-bit
stub in boot.asm also prints an A to the character buffer,
so that in the emulator I have a visual cue to the fact that everything
worked.

+++ bootstrap/src/c/include/elf.h
@@ -3,6 +3,6 @@
 
 #include "global.h"
 
-uint32_t *load_elf(uint32_t elf_addr);
+uint64_t *load_elf(uint32_t elf_addr);
 
 #endif
